set my view to be the home page
